+++ src/rpm/headers/types.rs
@@ -1,5 +1,8 @@
 //! A collection of types used in various header records.
+use std::str::FromStr;
+
 use crate::{constants::*, errors, Timestamp};
+use capctl::FileCaps;
 use digest::Digest;
 
 /// Offsets into an RPM Package (from the start of the file) demarking locations of each section
@@ -25,6 +28,7 @@ pub struct PackageFileEntry {
     pub group: String,
     pub base_name: String,
     pub dir: String,
+    pub caps: Option<FileCaps>,
     pub(crate) content: Vec<u8>,
 }
 
@@ -171,6 +175,8 @@ impl From<FileMode> for u16 {
 /// Description of file modes.
 ///
 /// A subset
+
+#[derive(Debug)]
 pub struct FileOptions {
     pub(crate) destination: String,
     pub(crate) user: String,
@@ -179,6 +185,7 @@ pub struct FileOptions {
     pub(crate) mode: FileMode,
     pub(crate) flag: FileFlags,
     pub(crate) inherit_permissions: bool,
+    pub(crate) caps: Option<FileCaps>,
 }
 
 impl FileOptions {
@@ -193,11 +200,13 @@ impl FileOptions {
                 mode: FileMode::regular(0o664),
                 flag: FileFlags::empty(),
                 inherit_permissions: true,
+                caps: None,
             },
         }
     }
 }
 
+#[derive(Debug)]
 pub struct FileOptionsBuilder {
     inner: FileOptions,
 }
@@ -222,6 +231,19 @@ impl FileOptionsBuilder {
         self.inner.mode = mode.into();
         self.inner.inherit_permissions = false;
         self
+    }
+
+    pub fn caps(mut self, caps: impl Into<String>) -> Result<Self, errors::Error> {
+        // verify capabilities
+        self.inner.caps = match FileCaps::from_str(&caps.into()) {
+            Ok(caps) => Some(caps),
+            Err(e) => {
+                return Err(errors::Error::InvalidCapabilities {
+                    caps: e.to_string(),
+                })
+            }
+        };
+        Ok(self)
     }
 
     pub fn is_doc(mut self) -> Self {
@@ -345,7 +367,6 @@ impl<W: std::io::Write> std::io::Write for Sha256Writer<W> {
 }
 
 mod test {
-
     #[test]
     fn test_file_mode() -> Result<(), Box<dyn std::error::Error>> {
         use super::*;
@@ -439,6 +460,19 @@ mod test {
             assert_eq!(raw_mode as u16, mode.raw_mode());
             assert_eq!(expected_type, mode.file_type());
         }
+        Ok(())
+    }
+
+    #[test]
+    fn test_verify_capabilities_valid() {
+        let blank_file = crate::FileOptions::new("/usr/bin/awesome");
+        blank_file.caps("cap_net_admin,cap_net_raw+p").unwrap();
+    }
+
+    #[test]
+    fn test_verify_capabilities_invalid() -> Result<(), crate::errors::Error> {
+        let blank_file = crate::FileOptions::new("/usr/bin/awesome");
+        blank_file.caps("cap_net_an,cap_net_raw+p").unwrap_err();
         Ok(())
     }
 }