+++ src/rpm/package.rs
@@ -832,11 +832,29 @@ impl PackageMetadata {
         let flags = self
             .header
             .get_entry_data_as_u32_array(IndexTag::RPMTAG_FILEFLAGS);
-        // @todo
-        // let caps = self.get_entry_i32_array_data(IndexTag::RPMTAG_FILECAPS)?;
 
-        match (modes, users, groups, digests, mtimes, sizes, flags) {
-            (Ok(modes), Ok(users), Ok(groups), Ok(digests), Ok(mtimes), Ok(sizes), Ok(flags)) => {
+        // Look for the file capabilities tag
+        // but it's not required so don't error out if it's not
+        let caps = match self
+            .header
+            .get_entry_data_as_string_array(IndexTag::RPMTAG_FILECAPS)
+        {
+            Ok(caps) => Ok(Some(caps)),
+            Err(Error::TagNotFound(_)) => Ok(None),
+            Err(e) => return Err(e),
+        };
+
+        match (modes, users, groups, digests, mtimes, sizes, flags, caps) {
+            (
+                Ok(modes),
+                Ok(users),
+                Ok(groups),
+                Ok(digests),
+                Ok(mtimes),
+                Ok(sizes),
+                Ok(flags),
+                Ok(caps),
+            ) => {
                 let paths = self.get_file_paths()?;
                 let n = paths.len();
 
@@ -850,13 +868,18 @@ impl PackageMetadata {
                     sizes,
                     flags,
                 ))
+                .enumerate()
                 .try_fold::<Vec<FileEntry>, _, Result<_, Error>>(
                     Vec::with_capacity(n),
-                    |mut acc, (path, user, group, mode, digest, mtime, size, flags)| {
+                    |mut acc, (idx, (path, user, group, mode, digest, mtime, size, flags))| {
                         let digest = if digest.is_empty() {
                             None
                         } else {
                             Some(FileDigest::load_from_str(algorithm, digest)?)
+                        };
+                        let cap = match caps {
+                            Some(caps) => caps.get(idx).map(|x| x.to_owned()),
+                            None => None,
                         };
                         acc.push(FileEntry {
                             path,
@@ -869,6 +892,7 @@ impl PackageMetadata {
                             digest,
                             flags: FileFlags::from_bits_retain(flags),
                             size: size as usize,
+                            caps: cap,
                         });
                         Ok(acc)
                     },
@@ -883,8 +907,9 @@ impl PackageMetadata {
                 Err(Error::TagNotFound(_)),
                 Err(Error::TagNotFound(_)),
                 Err(Error::TagNotFound(_)),
+                Err(Error::TagNotFound(_)),
             ) => Ok(vec![]),
-            (modes, users, groups, digests, mtimes, sizes, flags) => {
+            (modes, users, groups, digests, mtimes, sizes, flags, caps) => {
                 modes?;
                 users?;
                 groups?;
@@ -892,6 +917,7 @@ impl PackageMetadata {
                 mtimes?;
                 sizes?;
                 flags?;
+                caps?;
                 unreachable!()
             }
         }